Move real test utils out from testUtilities/fake
Merge test strategy into fake/serategy.ts
