+++ ironfish/src/blockchain/blockchain.test.ts
@@ -5,7 +5,7 @@
 import { Assert } from '../assert'
 import { AsyncUtils } from '../utils'
 import { createNodeTest, useAccountFixture, useBlockFixture } from '../testUtilities'
-import { makeBlockAfter, addBlocksShuffle } from '../testUtilities/fake'
+import { makeBlockAfter, addBlocksShuffle } from '../testUtilities/helpers/blockchain'
 
 describe('Blockchain', () => {
   const nodeTest = createNodeTest()
