+++ ironfish/src/testUtilities/fake/helpers/blockchain.ts
@@ -2,25 +2,24 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import { Account } from '../../../account'
 import { Assert } from '../../../assert'
-import { BlockSyncer } from '../../../blockSyncer'
 import { createRootLogger } from '../../../logger'
 import { GENESIS_BLOCK_PREVIOUS } from '../../../consensus'
 import { IDatabase } from '../../../storage'
-import { IronfishBlock, IronfishBlockchain, IronfishBlockHeader } from '../../../strategy'
-import { IronfishNode } from '../../../node'
-import { makeDb, makeDbName } from './storage'
+import { makeDb, makeDbName } from '../../helpers/storage'
 import { RangeHasher } from '../../../merkletree'
-import { SerializedTestTransaction, TestStrategy, TestTransaction } from '../strategy'
+import {
+  SerializedTestTransaction,
+  TestBlockchain,
+  TestStrategy,
+  TestTransaction,
+} from '../strategy'
 import { Spend } from '../../../strategy/transaction'
 import { Target } from '../../../blockchain/target'
-import { useBlockFixture, useMinerBlockFixture } from '../../fixtures'
 import Block from '../../../blockchain/block'
 import Blockchain from '../../../blockchain'
 import BlockHeader, { BlockHash } from '../../../blockchain/blockheader'
 import Strategy from '../../../strategy/strategy'
-import { MemPool } from '../../../memPool'
 import {
   BlockRequest,
   BlocksResponse,
@@ -29,54 +28,6 @@ import {
   NodeMessageType,
 } from '../../../network/messages'
 import { Direction } from '../../../network/messageRouters/rpc'
-
-/**
- * Type of a test anchorchain, encompassing the various generic parameters.
- */
-export type TestBlockchain = Blockchain<
-  string,
-  string,
-  TestTransaction,
-  string,
-  string,
-  SerializedTestTransaction
->
-
-export type TestMemPool = MemPool<
-  string,
-  string,
-  TestTransaction,
-  string,
-  string,
-  SerializedTestTransaction
->
-
-export type TestBlockSyncer = BlockSyncer<
-  string,
-  string,
-  TestTransaction,
-  string,
-  string,
-  SerializedTestTransaction
->
-
-export type TestBlockHeader = BlockHeader<
-  string,
-  string,
-  TestTransaction,
-  string,
-  string,
-  SerializedTestTransaction
->
-
-export type TestBlock = Block<
-  string,
-  string,
-  TestTransaction,
-  string,
-  string,
-  SerializedTestTransaction
->
 
 /**
  * Add the notes directly to the anchorchain's notes merkle tree
@@ -321,135 +272,6 @@ export async function makeChainFull(
   }
 
   return chain
-}
-
-/**
- * Create a test block with no transactions, that can be after any block either on the chain or not.
- * It works by not affecting the merkle trees at all, and requires this block to have no transactions,
- * therefore no notes.
- *
- * @param chain the chain is not used, only the verifier and strategy from the chain
- * @param after the block the created block should be after
- */
-export async function makeBlockAfter(
-  chain: IronfishBlockchain,
-  after: IronfishBlockHeader | IronfishBlock,
-): Promise<IronfishBlock> {
-  if (after instanceof Block) {
-    after = after.header
-  }
-
-  const sequence = after.sequence + BigInt(1)
-  const miningReward = BigInt(chain.strategy.miningReward(sequence))
-
-  if (miningReward !== BigInt(0)) {
-    throw new Error(`Must have mining reward disabled but was ${miningReward}`)
-  }
-
-  const timestamp = new Date()
-  const target = Target.calculateTarget(timestamp, after.timestamp, after.target)
-  const randomness = Math.random()
-  const graffiti = Buffer.alloc(32)
-  graffiti.write('fake block')
-
-  const header = new BlockHeader(
-    chain.strategy,
-    sequence,
-    after.hash,
-    after.noteCommitment,
-    after.nullifierCommitment,
-    target,
-    randomness,
-    timestamp,
-    miningReward,
-    graffiti,
-    true,
-    BigInt(1),
-  )
-
-  const block = new Block(header, [])
-
-  Assert.isTrue((await chain.verifier.verifyBlock(block)).valid === 1)
-  return block
-}
-
-/**
- * Adds a block to the chain that gives {@link from} a
- * miners fee, then a transaction on a new block that
- * gives that miners fee to {@link to}, as well as another
- * miners fee for {@link from}.
- *
- * Returned block has 1 spend, 3 notes
- */
-export async function makeBlockWithTransaction(
-  node: IronfishNode,
-  from: Account,
-  to: Account,
-): Promise<IronfishBlock> {
-  const head = await node.chain.getHeaviestHead()
-  Assert.isNotNull(head, 'No genesis block. Call node.seed() first')
-  const sequence = head.sequence
-
-  const block1 = await useMinerBlockFixture(
-    node.chain,
-    sequence + BigInt(1),
-    from,
-    node.accounts,
-  )
-
-  await node.chain.addBlock(block1)
-  await node.accounts.updateHead()
-
-  const block2 = await useBlockFixture(node.chain, async () => {
-    const transaction = await node.accounts.createTransaction(
-      from,
-      BigInt(1),
-      BigInt(0),
-      '',
-      to.publicAddress,
-    )
-
-    return node.chain.newBlock(
-      [transaction],
-      await node.chain.strategy.createMinersFee(
-        await transaction.transactionFee(),
-        sequence + BigInt(2),
-        from.spendingKey,
-      ),
-    )
-  })
-
-  return block2
-}
-
-/**
- * This adds blocks to a chain in random order. It's useful to help root out bugs where insertion order
- * can create bugs because someone accidently wrote code that is graph structure dependent. If any block
- * fails to be added, the operation will stop and return false
- *
- * @param chain the chain to insert blocks into
- * @param blocks the blocks to insert in random order
- * @param randomDrop should it randomly decide drop blocks with a 10% chance
- */
-export async function addBlocksShuffle(
-  chain: IronfishBlockchain,
-  blocks: IronfishBlock[],
-  randomDrop = false,
-): Promise<boolean> {
-  blocks = [...blocks]
-
-  while (blocks.length > 0) {
-    const index = Math.floor(Math.random() * blocks.length)
-    const block = blocks.splice(index, 1)[0]
-
-    const shouldDrop = randomDrop && Math.random() > 0.9
-    if (shouldDrop) continue
-
-    const { isAdded } = await chain.addBlock(block)
-    if (!isAdded) return false
-  }
-
-  return true
 }
 
 /**
