--- ironfish/src/testUtilities/fake/strategy/TestStrategy.ts
@@ -1,104 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-
-import { BlockHash } from '../../../blockchain/blockheader'
-import { NullifierHasher } from '../../../blockchain/nullifiers'
-import Strategy from '../../../strategy/strategy'
-import { ConcatHasher } from '../../../merkletree'
-
-import { TestTransaction } from './TestTransaction'
-import { TestTransactionSerde } from './TestTransactionSerde'
-import { SerializedTestTransaction } from './SerializedTypes'
-import { TestVerifier } from './testVerifier'
-import { TestBlockchain } from '../helpers'
-import { BlockSerde } from '../../../blockchain'
-
-/**
- * Very basic strategy for testing blockchain code. Models notes and hashes
- * as concatenated strings, and uses dumb calculations for hashing and
- * target calculations
- */
-export class TestStrategy
-  implements
-    Strategy<string, string, TestTransaction, string, string, SerializedTestTransaction> {
-  _noteHasher: ConcatHasher
-  _nullifierHasher: NullifierHasher
-
-  _blockSerde: BlockSerde<
-    string,
-    string,
-    TestTransaction<string>,
-    string,
-    string,
-    SerializedTestTransaction<string>
-  >
-
-  constructor(noteHasher = new ConcatHasher()) {
-    this._noteHasher = noteHasher
-    this._nullifierHasher = new NullifierHasher()
-    this._blockSerde = new BlockSerde(this)
-  }
-
-  createVerifier(chain: TestBlockchain): TestVerifier {
-    return new TestVerifier(chain)
-  }
-
-  noteHasher(): ConcatHasher {
-    return this._noteHasher
-  }
-
-  nullifierHasher(): NullifierHasher {
-    return this._nullifierHasher
-  }
-
-  transactionSerde(): TestTransactionSerde {
-    return new TestTransactionSerde()
-  }
-
-  get blockSerde(): BlockSerde<
-    string,
-    string,
-    TestTransaction<string>,
-    string,
-    string,
-    SerializedTestTransaction<string>
-  > {
-    return this._blockSerde
-  }
-
-  /**
-   * Generate a hash from the block's sequence.
-   */
-  hashBlockHeader(serializedHeader: Buffer): BlockHash {
-    const headerWithoutRandomness = Buffer.from(serializedHeader.slice(8))
-    const header = JSON.parse(headerWithoutRandomness.toString()) as Record<string, unknown>
-
-    const sequence = BigInt(header['sequence'])
-    const bigIntArray = BigInt64Array.from([sequence])
-    const byteArray = Buffer.from(bigIntArray.buffer)
-    const result = Buffer.alloc(32)
-    result.set(byteArray)
-    return result
-  }
-
-  createMinersFee(
-    totalTransactionFees: bigint,
-    blockSequence: bigint,
-    _minerKey: string,
-  ): Promise<TestTransaction> {
-    const miningReward = this.miningReward(blockSequence)
-    return Promise.resolve(
-      new TestTransaction(
-        true,
-        [`miners note ${totalTransactionFees + BigInt(miningReward)}`],
-        BigInt(-1) * (totalTransactionFees + BigInt(miningReward)),
-        [],
-      ),
-    )
-  }
-
-  miningReward(_blockSequence: bigint): number {
-    return 10
-  }
-}
