--- ironfish/src/testUtilities/fake/strategy/index.ts
@@ -1,8 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-
-export * from './SerializedTypes'
-export * from './TestStrategy'
-export * from './TestTransaction'
-export * from './TestTransactionSerde'
